Changed Priority from Int to Double for more precise calculations; Added Migration for updated Wishes Table;

+++ app/src/main/java/com/sar/shopaholism/data/local/entity/WishEntity.kt
@@ -23,5 +23,5 @@ data class WishEntity(
     val price: Double,
 
     @ColumnInfo(name = "priority")
-    val priority: Int
+    val priority: Double
 )
